Rename SiriDomain name to be SiriShortcuts.Domain and SiriShortcuts.Suggestions to reduce global name pollution.
